move utils/loss.py to models/functions/loss.py
