+++ tests/models/functions/test_loss.py
@@ -1,8 +1,8 @@
 import pytest
 import torch
 
+from kwja.models.functions.loss import compute_multi_label_token_mean_loss, compute_token_mean_loss, mask_logits
 from kwja.utils.constants import MASKED
-from kwja.utils.loss import compute_multi_label_token_mean_loss, compute_token_mean_loss, mask_logits
 
 
 @pytest.mark.parametrize(
